+++ spec/active_record_consumer_spec.rb
@@ -61,10 +61,10 @@ module ActiveRecordConsumerTest
     end
 
     describe 'consume' do
-      SCHEMA_CLASS_SETTINGS.each do |setting, use_schema_class|
+      SCHEMA_CLASS_SETTINGS.each do |setting, use_schema_classes|
         context "with Schema Class consumption #{setting}" do
           before(:each) do
-            Deimos.configure { |config| config.schema.use_schema_class = use_schema_class }
+            Deimos.configure { |config| config.schema.use_schema_classes = use_schema_classes }
           end
 
           it 'should receive events correctly' do