+++ spec/batch_consumer_spec.rb
@@ -30,10 +30,10 @@ module ConsumerTest
     end
 
     describe 'consume_batch' do
-      SCHEMA_CLASS_SETTINGS.each do |setting, use_schema_class|
+      SCHEMA_CLASS_SETTINGS.each do |setting, use_schema_classes|
         context "with Schema Class consumption #{setting}" do
           before(:each) do
-            Deimos.configure { |config| config.schema.use_schema_class = use_schema_class }
+            Deimos.configure { |config| config.schema.use_schema_classes = use_schema_classes }
           end
 
           it 'should provide backwards compatibility for BatchConsumer class' do