+++ spec/config/configuration_spec.rb
@@ -21,7 +21,7 @@ describe Deimos, 'configuration' do
       kafka_logger logger
       reraise_consumer_errors true
       schema_registry_url 'http://schema.registry'
-      schema.use_schema_class false
+      schema.use_schema_classes false
       seed_broker 'whatever'
       schema_path 'some_path'
       producer_schema_namespace 'namespace'
@@ -38,7 +38,7 @@ describe Deimos, 'configuration' do
     expect(described_class.config.kafka.logger).to eq(logger)
     expect(described_class.config.consumers.reraise_errors).to eq(true)
     expect(described_class.config.schema.registry_url).to eq('http://schema.registry')
-    expect(described_class.config.schema.use_schema_class).to eq(false)
+    expect(described_class.config.schema.use_schema_classes).to eq(false)
     expect(described_class.config.kafka.seed_brokers).to eq('whatever')
     expect(described_class.config.producers.schema_namespace).to eq('namespace')
     expect(described_class.config.producers.topic_prefix).to eq('prefix')
@@ -89,7 +89,7 @@ describe Deimos, 'configuration' do
           offset_retention_time: nil,
           heartbeat_interval: 10,
           handler: 'ConsumerTest::MyConsumer',
-          use_schema_class: nil
+          use_schema_classes: nil
         }, {
           topic: 'my_batch_consume_topic',
           group_id: 'my_batch_group_id',
@@ -106,7 +106,7 @@ describe Deimos, 'configuration' do
           offset_retention_time: nil,
           heartbeat_interval: 10,
           handler: 'ConsumerTest::MyBatchConsumer',
-          use_schema_class: nil
+          use_schema_classes: nil
         }
       ],
       producer: {
@@ -182,7 +182,7 @@ describe Deimos, 'configuration' do
         offset_commit_threshold 13
         offset_retention_time 13
         heartbeat_interval 13
-        use_schema_class false
+        use_schema_classes false
       end
       consumer do
         disabled true
@@ -190,7 +190,7 @@ describe Deimos, 'configuration' do
         schema 'blah2'
         topic 'blah2'
         group_id 'myconsumerid2'
-        use_schema_class false
+        use_schema_classes false
       end
     end
 
@@ -234,7 +234,7 @@ describe Deimos, 'configuration' do
             offset_retention_time: 13,
             heartbeat_interval: 13,
             handler: 'MyConfigConsumer',
-            use_schema_class: false
+            use_schema_classes: false
           }
         ],
         producer: {