+++ spec/schema_classes/my_nested_schema.rb
@@ -19,7 +19,6 @@ module Deimos
 
     # @override
     def initialize(**kwargs)
-      super()
       args = kwargs.with_indifferent_access
       @test_id = args[:test_id]
       @test_float = args[:test_float]