+++ spec/schema_classes/my_schema_with_complex_types.rb
@@ -25,7 +25,6 @@ module Deimos
 
     # @override
     def initialize(**kwargs)
-      super()
       args = kwargs.with_indifferent_access
       @test_id = args[:test_id]
       @test_float = args[:test_float]