+++ CHANGELOG.md
@@ -10,8 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features :star:
 
 - Generate Schema classes from Avro Schemas
-- Use Schema Classes in your consumer and producer
-- Guardfile to re-run `rake deimos:generate_schema_models` upon changes to your schemas 
+- Use Schema Classes in your consumer and producer 
 
 ## 1.11.0 - 2021-08-27